OneDrive/Documentos/GSI/Super Paqui/whatsapp/tablero/requirements.txt
fastapi
uvicorn
jinja2
gspread
google-auth
python-multipart
